% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cardscount.R
\name{calculate_blackjack_score}
\alias{calculate_blackjack_score}
\title{Calculate the score of a hand in blackjack
description
Calculate the score of a hand in blackjack based on the numeric value of the cards and the rules of the game.
The function computes the score by adding up the numeric value of number cards,
treating face cards (J, Q, K) as 10, and handling aces (A) based on whether
adding 11 would result in a bust (score > 21).}
\usage{
calculate_blackjack_score(cards)
}
\arguments{
\item{cards}{A character vector of cards (e.g., "2", "J", "A")}
}
\value{
The total score of the hand
}
\description{
Calculate the score of a hand in blackjack
description
Calculate the score of a hand in blackjack based on the numeric value of the cards and the rules of the game.
The function computes the score by adding up the numeric value of number cards,
treating face cards (J, Q, K) as 10, and handling aces (A) based on whether
adding 11 would result in a bust (score > 21).
}
\details{
The function computes the score by adding up the numeric value of number cards,
treating face cards (J, Q, K) as 10, and handling aces (A) based on whether
adding 11 would result in a bust (score > 21).
}
\examples{
# calculate_blackjack_score(c("2", "J", "A"))

}
