% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cardscount-package.R
\docType{package}
\name{Cardscount-package}
\alias{Cardscount}
\alias{Cardscount-package}
\title{Cardscount: A Simple Blackjack Score Calculator}
\description{
A package to calculate the score of a hand in blackjack.
}
\author{
\strong{Maintainer}: Hai Phan \email{ppha0028@student.monash.edu}

}
\keyword{internal}
