% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cardscount.R
\name{play_blackjack}
\alias{play_blackjack}
\title{play_blackjack}
\usage{
play_blackjack()
}
\value{
The total score of the hand
}
\description{
Calculate the score of a hand in blackjack
}
\details{
The function computes the score by adding up the numeric value of number cards,
treating face cards (J, Q, K) as 10, and handling aces (A) based on whether
adding 11 would result in a bust (score > 21).
}
\examples{
# play_blackjack()
}
